checkpoint: issue with localeswitcher component
