+++ dictionaries/it.json
@@ -1,0 +1,3 @@
+{
+    "test": "ciao"
+}
