+++ i18n.config.ts
@@ -1,0 +1,6 @@
+export const i18n = {
+    defaultLocale: "it",
+    locales: ["it", "en"],
+} as const;
+  
+export type Locale = (typeof i18n)["locales"][number];
