+++ package.json
@@ -15,6 +15,7 @@
     "add-to-calendar-button-react": "^2.6.10",
     "autoprefixer": "^10.4.19",
     "firebase": "^10.9.0",
+    "flag-icons": "^7.2.1",
     "gray-matter": "^4.0.3",
     "image-size": "^1.1.1",
     "luxon": "^3.4.4",
@@ -32,19 +33,19 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/image-size": "^0.8.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.6",
-    "@typescript-eslint/parser": "^5.59.6",
-    "@vitejs/plugin-basic-ssl": "^1.1.0",
-    "@vitejs/plugin-react": "^4.2.1",
     "@headlessui/react": "^1.7.18",
     "@heroicons/react": "^2.1.3",
     "@tailwindcss/typography": "^0.5.12",
+    "@types/image-size": "^0.8.0",
     "@types/luxon": "^3.4.2",
     "@types/node": "^20.12.5",
     "@types/react": "18.2.75",
     "@types/react-dom": "18.2.22",
     "@types/react-helmet": "^6.1.11",
+    "@typescript-eslint/eslint-plugin": "^5.59.6",
+    "@typescript-eslint/parser": "^5.59.6",
+    "@vitejs/plugin-basic-ssl": "^1.1.0",
+    "@vitejs/plugin-react": "^4.2.1",
     "dotenv-cli": "^7.4.1",
     "eslint": "8.57.0",
     "eslint-config-prettier": "^9.1.0",
