+++ src/pages/_app.tsx
@@ -1,4 +1,5 @@
 import 'tailwindcss/tailwind.css';
+import '@/globals.css'
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 import React from 'react';
