+++ src/pages/_document.tsx
@@ -5,7 +5,7 @@ export default function Document() {
   return (
     <Html>
       <Head />
-      <body className='dark:bg-slate-900 dark:text-slate-300' lang='it'>
+      <body className='dark:bg-slate-900 dark:text-slate-300'>
         <Main />
         <NextScript />
       </body>
