+++ src/utils/locale.ts
@@ -1,0 +1,9 @@
+import { i18n } from "i18n.config"
+
+export const getDefaultLocale = () => {
+    return i18n.defaultLocale
+}
+
+export const getAllLocales = () => {
+    return i18n.locales
+}
